Refactor part of web plugin into common logic. (#5807)
* feat: modify generateHtml to generateEntry

* feat: support entry

* chore: update lock

* feat: support extra script

* feat: add config for entry

* feat: support entryType

* feat: add documentToJs

* feat: remove ejs to template

* chore: modify entry to output

* feat: add dom render

* chore: update lock

* chore: should publish templates

* feat: modify type and merge config

* feat: remove setConfig for output

* chore: modify jsOutput

* feat: add paths of outputs

* fix: should get data when after.start.compile

* feat: support dev

* feat: modify dist type

* chore: modify htmlOutput

* fix: modify dirname

* chore: modify dirname

* chore: add type for __ICE__CREATE_ELEMENT

* test: add test for domRender

* feat: add renderToHTML

* chore: modify platform

* feat: support target and renderMode

* fix: test case and renderer

* chore: update lock

* Update build.ts

* chore: add default value to urls

* refactor: constant

* feat: add with weex

* chore: change multi target

* fix: output orders

* fix: build order

* feat: modify disttype

* feat: modify textContent to support styleElement in weex

* fix: dist type javascript

* chore: version change

* fix: merge and versions

* fix: target renderer

* chore: remove with weex example

* fix: add commandArgs defaults for target, rename miniapp platforms to targets

* fix: bundle slash
