+++ packages/ice/bin/ice-cli.mjs
@@ -26,7 +26,7 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
     .description('build project')
     .allowUnknownOption()
     .addOption(new Option('--target <target>', 'set build target').default(WEB).choices(TARGETS))
-    .option('--platform <platform>', 'same as --target', WEB)
+    .option('--target <target>', 'same as --target', WEB)
     .option('--mode <mode>', 'set mode', 'production')
     .option('--analyzer', 'visualize size of output files', false)
     .option('--config <config>', 'use custom config')
@@ -43,7 +43,7 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
     .description('start server')
     .allowUnknownOption()
     .addOption(new Option('--target <target>', 'set build target').default(WEB).choices(TARGETS))
-    .option('--platform <platform>', 'same as --target', WEB)
+    .option('--target <target>', 'same as --target', WEB)
     .option('--mode <mode>', 'set mode', 'development')
     .option('--config <config>', 'custom config path')
     .option('-h, --host <host>', 'dev server host', '0.0.0.0')
