+++ packages/ice/src/plugins/web/task.ts
@@ -44,6 +44,8 @@ const getWebTask = ({ rootDir, command, dataCache, userConfig }): Config => {
     fastRefresh: command === 'start',
     logging: process.env.WEBPACK_LOGGING || defaultLogging,
     minify: command === 'build',
+    useDevServer: true,
+    useDataLoader: true,
   };
 };
