+++ packages/plugin-miniapp/src/targets/ali/index.ts
