+++ packages/plugin-miniapp/src/targets/ali/template.ts
