+++ packages/plugin-miniapp/src/targets/baidu/index.ts
