+++ packages/plugin-miniapp/src/targets/baidu/template.ts
