+++ packages/plugin-miniapp/src/targets/bytedance/index.ts
