+++ packages/plugin-miniapp/src/targets/bytedance/template.ts
