+++ packages/plugin-miniapp/src/targets/wechat/index.ts
