+++ packages/plugin-miniapp/src/targets/wechat/template.ts
